update aiounittest dependency in light of 3.8 features
The aiounittest docs note that python 3.8 has gained the
unittest.IsolatedAsyncioTestCase class, making the library obsolete.
Because of this, the library is now only required on 3.7 and below.
The import statement in the test file has been updated for both cases.
Also added an empty init file so we can just run python -m unittest.